+++ src/main/java/utils/Vector2.java
@@ -5,22 +5,22 @@ import java.io.Serializable;
 public class Vector2 implements Serializable
 {
 	private static final long serialVersionUID = 8642244552320036511L;
-	private double x;
-	private double y;
+	private float x;
+	private float y;
 	
-	public Vector2(double x, double y)
+	public Vector2(float x, float y)
 	{
 		this.setX(x);
 		this.setY(y);
 	}
 	
-	public double len2()
+	public float len2()
 	{
 		return x*x + y*y;
 	}
 	
-	public double len(){
-		return Math.sqrt(x*x + y*y);
+	public float len(){
+		return (float) Math.sqrt(x*x + y*y);
 	}
 
 	public Vector2 copy() {
@@ -33,7 +33,7 @@ public class Vector2 implements Serializable
 		return this;
 	}
 
-	public Vector2 scale(double s) {
+	public Vector2 scale(float s) {
 		x *= s;
 		y *= s;
 		return this;
@@ -47,11 +47,11 @@ public class Vector2 implements Serializable
 		return new Vector2(getX() - b.getX(), getY() - b.getY());
 	}
 	
-	public Vector2 mul(double s) {
+	public Vector2 mul(float s) {
 		return new Vector2(s*getX(), s*getY());
 	}
 	
-	public double dot(Vector2 b) {
+	public float dot(Vector2 b) {
 		return ( getX()*b.getX() + getY()*b.getY() );
 	}
 	
@@ -59,44 +59,44 @@ public class Vector2 implements Serializable
 		return new Vector2(-getY(), getX());
 	}
 	
-	public Vector2 rotate(double angle)
+	public Vector2 rotate(float angle)
 	{
-		double c = Math.cos(angle);
-		double s = Math.sin(angle);
+		float c = (float) Math.cos(angle);
+		float s = (float) Math.sin(angle);
 		return new Vector2(x*c - y*s, x*s + y*c);
 	}
 	
 	public Vector2 unit()
 	{
-		double len = len();
+		float len = len();
 		if (len == 0)
 			return new Vector2(0, 0);
 		return new Vector2(getX() / len, getY() / len);
 	}
 	
-	public double angle() {
-		return Math.atan2(getY(), getX());
+	public float angle() {
+		return (float) Math.atan2(getY(), getX());
 	}
 
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 
-	public Vector2 setLength(double targetLen)
+	public Vector2 setLength(float targetLen)
 	{
-		double currentLen = len();
+		float currentLen = len();
 		x *= targetLen / currentLen;
 		y *= targetLen / currentLen;
 		return this;
@@ -112,21 +112,21 @@ public class Vector2 implements Serializable
 		return "(" + x + ", " + y + ")";
 	}
 
-	public double angleBetween(Vector2 other)
+	public float angleBetween(Vector2 other)
 	{
-		double a = len2(), b = other.len2();
-		return Math.acos(dot(other) / Math.sqrt(a*b));
+		float a = len2(), b = other.len2();
+		return (float) Math.acos(dot(other) / Math.sqrt(a*b));
 	}
 
-	public double distanceTo(Vector2 other) {
-		double dx = getX() - other.getX();
-		double dy = getY() - other.getY();
-		return Math.sqrt(dx*dx + dy*dy);
+	public float distanceTo(Vector2 other) {
+		float dx = getX() - other.getX();
+		float dy = getY() - other.getY();
+		return (float) Math.sqrt(dx*dx + dy*dy);
 	}
 
-	public double squareDistanceTo(Vector2 other) {
-		double dx = getX() - other.getX();
-		double dy = getY() - other.getY();
+	public float squareDistanceTo(Vector2 other) {
+		float dx = getX() - other.getX();
+		float dy = getY() - other.getY();
 		return dx*dx + dy*dy;
 	}
 
@@ -138,10 +138,10 @@ public class Vector2 implements Serializable
 		return false;
 	}
 
-	public void moveAway(Vector2 other, double amount) {
-		double dx = getX() - other.getX();
-		double dy = getY() - other.getY();
-		double l = Math.sqrt(dx*dx + dy*dy);
+	public void moveAway(Vector2 other, float amount) {
+		float dx = getX() - other.getX();
+		float dy = getY() - other.getY();
+		float l = (float) Math.sqrt(dx*dx + dy*dy);
 		x = other.getX() + dx * amount / l;
 		y = other.getY() + dy * amount / l;
 	}
